some clean-up
