+++ client/pages/_app.tsx
@@ -1,7 +1,7 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
 import { ChakraProvider } from "@chakra-ui/react";
-import { AuthProvider } from "../context/AuthContext2";
+import { AuthProvider } from "../context/AuthContext";
 import { customTheme } from "../themes/index";
 
 function MyApp({ Component, pageProps }: AppProps) {
